fix on button, text and combobox for eraser
